fix rounding, transaction safety, and FIFO determinism bugs
- Promote rounding_div to i128 params to prevent overflow on large
  BTC amounts; fix negative rounding behavior (half-away-from-zero)
  and add div-by-zero assert
- Wrap allocate command in DB transaction with AllocateError; fix
  satoshis proportional split using rounding_div instead of direct
  division; copy acquisition_fair_values links when splitting lots
- Add FIFO tiebreaker (id ASC) to import, transfer, and allocate
  to ensure deterministic lot ordering
- Deduplicate date parsing: extract parse_date_str in models, reuse
  in transfer.rs and deserialize_date
- Fix IRS long-term threshold: > 365 days, not >= 365
- Move load_lot_config() inside Import arm only

+++ src/commands/report.rs
@@ -138,7 +138,7 @@ pub fn report_gaap_term(wtr: &mut Writer<File>, beg: NaiveDateTime, end: NaiveDa
         let basis = (Decimal::from_i64(acq_disp.2.basis).unwrap() / dec!(100)).round_dp_with_strategy(2, RoundingStrategy::MidpointAwayFromZero);
 
         // Cost basis for the disposed sats (original acquisition price)
-        let cost_basis_cents = rounding_div(acq_disp.2.satoshis * acq_disp.1.usd_cents_btc_basis, 100_000_000);
+        let cost_basis_cents = rounding_div(acq_disp.2.satoshis as i128 * acq_disp.1.usd_cents_btc_basis as i128, 100_000_000);
         let cost_basis = (Decimal::from_i64(cost_basis_cents).unwrap() / dec!(100)).round_dp_with_strategy(2, RoundingStrategy::MidpointAwayFromZero);
 
         // FMV Disposed = fair value basis - cost basis for same sats
